requirements.txt
httpx>=0.27
rich>=13.7
typer>=0.12
PyYAML>=6.0
